+++ apollo-federation-types/CHANGELOG.md
@@ -2,15 +2,21 @@
 
 Not every version is listed here because versions before 0.14.0 did not have a changelog.
 
-## 0.15.0 (Unreleased)
+## 0.15.0 - Unreleased
 
 ### Breaking changes
 
 - `GraphQLError.nodes` is now an `Option<Vec<SubgraphASTNode>>`
+- All usages of `camino::Utf8PathBuf` have been replaced with `std::path::PathBuf`
+
+### Features
+
+- A new `json_schema` feature derives the `schemars::JsonSchema` trait on `SupergraphConfig` and its sub-types.
 
 ## 0.14.1 - 2024-09-19
 
 ### Features
+
 - `impl FromIterator<(String, SubgraphConfig)> for SupergraphConfig`
 
 ## 0.14.0 - 2024-09-11